feat(ui): add swap flow with processing, result screens, and cross-fade transitions
Refactor hero-right-sidebar into wallet-sidebar directory with separate
components. Add complete swap flow with form, processing animation,
success/error result screens, and transaction details view. Implement
content-only cross-fade transitions for all phase and tab changes.
