+++ frontend/src/components/wallet-sidebar/send-content.tsx
@@ -1,0 +1,27 @@
+"use client";
+
+export function SendContent() {
+  return (
+    <div
+      style={{
+        flex: 1,
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "center",
+        padding: "20px",
+      }}
+    >
+      <span
+        style={{
+          fontFamily: "var(--font-geist-sans), sans-serif",
+          fontSize: "16px",
+          fontWeight: 400,
+          lineHeight: "20px",
+          color: "rgba(60, 60, 67, 0.6)",
+        }}
+      >
+        Send — coming soon
+      </span>
+    </div>
+  );
+}
