+++ frontend/src/components/wallet-sidebar/shared.tsx
@@ -1,0 +1,135 @@
+"use client";
+
+import { ArrowLeft, Search, X } from "lucide-react";
+
+export function SearchInput({
+  value,
+  onChange,
+  placeholder = "Search",
+}: {
+  value: string;
+  onChange: (v: string) => void;
+  placeholder?: string;
+}) {
+  return (
+    <div style={{ padding: "8px 20px" }}>
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          background: "rgba(0, 0, 0, 0.04)",
+          borderRadius: "47px",
+          padding: "0 16px",
+          gap: "8px",
+          height: "44px",
+        }}
+      >
+        <Search
+          size={24}
+          style={{ color: "rgba(60, 60, 67, 0.6)", flexShrink: 0 }}
+        />
+        <input
+          onChange={(e) => onChange(e.target.value)}
+          placeholder={placeholder}
+          style={{
+            flex: 1,
+            background: "none",
+            border: "none",
+            outline: "none",
+            fontFamily: "var(--font-geist-sans), sans-serif",
+            fontSize: "16px",
+            fontWeight: 400,
+            lineHeight: "20px",
+            color: "#000",
+            padding: 0,
+          }}
+          type="text"
+          value={value}
+        />
+      </div>
+    </div>
+  );
+}
+
+export function SubViewHeader({
+  title,
+  onBack,
+  onClose,
+}: {
+  title: string;
+  onBack: () => void;
+  onClose: () => void;
+}) {
+  return (
+    <>
+      <style jsx>{`
+        .subview-back:hover {
+          background: rgba(0, 0, 0, 0.08) !important;
+        }
+        .subview-close:hover {
+          background: rgba(0, 0, 0, 0.08) !important;
+        }
+      `}</style>
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "space-between",
+          padding: "8px",
+        }}
+      >
+        <button
+          className="subview-back"
+          onClick={onBack}
+          style={{
+            width: "36px",
+            height: "36px",
+            display: "flex",
+            justifyContent: "center",
+            alignItems: "center",
+            background: "rgba(0, 0, 0, 0.04)",
+            border: "none",
+            borderRadius: "9999px",
+            cursor: "pointer",
+            transition: "all 0.2s ease",
+            color: "#3C3C43",
+          }}
+          type="button"
+        >
+          <ArrowLeft size={24} />
+        </button>
+        <span
+          style={{
+            fontFamily: "var(--font-geist-sans), sans-serif",
+            fontSize: "18px",
+            fontWeight: 600,
+            lineHeight: "28px",
+            color: "#000",
+          }}
+        >
+          {title}
+        </span>
+        <button
+          className="subview-close"
+          onClick={onClose}
+          style={{
+            width: "36px",
+            height: "36px",
+            display: "flex",
+            justifyContent: "center",
+            alignItems: "center",
+            background: "rgba(0, 0, 0, 0.04)",
+            border: "none",
+            borderRadius: "9999px",
+            cursor: "pointer",
+            transition: "all 0.2s ease",
+            color: "#3C3C43",
+          }}
+          type="button"
+        >
+          <X size={24} />
+        </button>
+      </div>
+    </>
+  );
+}
